[offchain-protocol] Move all schemas to @celo/payment-types package
