+++ package.json
@@ -11,11 +11,12 @@
   ],
   "scripts": {
     "clean": "rm -rf node_modules packages/**/node_modules packages/**/lib packages/**/build packages/**/tsconfig.tsbuildinfo",
+    "clean:build": "rm -rf packages/**/tsconfig.tsbuildinfo",
     "build": "lerna run build",
     "postinstall": "lerna bootstrap",
     "start": "yarn lerna run start --stream",
     "cli": "yarn build && yarn --cwd ./packages/cli cli",
-    "generate-web-models": "yarn --cwd ./packages/sdk web-models:build"
+    "generate-web-models": "yarn --cwd ./packages/types web-models:build"
   },
   "engines": {
     "node": ">=12"
@@ -31,8 +32,5 @@
   },
   "resolutions": {
     "@celo/utils": "1.2.0"
-  },
-  "devDependencies": {
-    "openapi-typescript-codegen": "^0.9.3"
   }
 }
