+++ packages/sdk/package.json
@@ -35,8 +35,7 @@
     "doc:json": "typedoc src/ --exclude **/*.spec.ts --target ES6 --mode file --json build/docs/typedoc.json",
     "doc:publish": "gh-pages -m \"[ci skip] Updates\" -d build/docs",
     "version": "standard-version",
-    "prepare-release": "run-s reset-hard test cov:check doc:html version doc:publish",
-    "web-models:build": "ts-node scripts/genereate-web-models.ts"
+    "prepare-release": "run-s reset-hard test cov:check doc:html version doc:publish"
   },
   "engines": {
     "node": ">=12"
