--- packages/sdk/src/schemas.ts
@@ -1,2 +1,0 @@
-export { GetPaymentInfoResponseData as PaymentInfo } from './offchain-protocol/models/GetPaymentInfoResponseData';
-export { PayerData } from './offchain-protocol/models/PayerData';
