+++ packages/types/package.json
@@ -7,9 +7,14 @@
     "lib"
   ],
   "scripts": {
-    "build": "tsc -b ."
+    "build": "tsc -b .",
+    "web-models:build": "ts-node ./scripts/genereate-web-models.ts"
   },
   "dependencies": {
     "@celo/contractkit": "1.2.0"
+  },
+  "devDependencies": {
+    "openapi-typescript-codegen": "^0.9.3",
+    "ts-node": "^10.2.1"
   }
 }
