+++ packages/types/src/logo.png
