+++ packages/types/src/schemas/GetPaymentInfoResponse.ts
