+++ packages/types/src/schemas/GetPaymentInfoResponseData.ts
