+++ packages/types/src/schemas/JsonRpcError.ts
