+++ packages/types/src/schemas/JsonRpcErrorResponse.ts
