+++ packages/types/src/schemas/JsonRpcInvalidParameterError.ts
