+++ packages/types/src/schemas/JsonRpcMethodNotFoundError.ts
