+++ packages/types/src/schemas/JsonRpcProtocol.ts
