--- packages/types/src/schemas/confirm.ts
@@ -1,8 +1,0 @@
-import { JsonRpcMethods } from "../methods";
-
-export interface Confirm {
-  method: JsonRpcMethods.Confirm;
-  params: {
-    referenceId: string;
-  };
-}
