+++ packages/types/src/schemas/index.ts
@@ -1,5 +1,27 @@
-export * from "./abort";
-export * from "./get-info";
-export * from "./init-charge";
-export * from "./confirm";
-export * from "./kyc";
+export * from "./Address";
+export * from "./BusinessData";
+export * from "./GetPaymentInfo";
+export * from "./GetPaymentInfoParams";
+export * from "./GetPaymentInfoResponse";
+export * from "./GetPaymentInfoResponseData";
+export * from "./InitCharge";
+export * from "./InitChargeParams";
+export * from "./InitChargeResponse";
+export * from "./JsonRpcError";
+export * from "./JsonRpcErrorResponse";
+export * from "./JsonRpcInvalidParameterError";
+export * from "./JsonRpcMethodNotFoundError";
+export * from "./JsonRpcProtocol";
+export * from "./JsonRpcRequired";
+export * from "./NationalIdData";
+export * from "./PayerData";
+export * from "./PaymentAction";
+export * from "./PaymentSender";
+export * from "./ReadyForSettlement";
+export * from "./ReadyForSettlementParams";
+export * from "./ReadyForSettlementResponse";
+export * from "./ReceiverData";
+export * from "./ReferenceId";
+export * from "./RequiredAddressData";
+export * from "./RequiredNationalIdData";
+export * from "./RequiredPayerData";
