--- packages/types/src/schemas/init-charge.ts
@@ -1,14 +1,0 @@
-import { JsonRpcMethods } from "../methods";
-import { KYC } from "./kyc";
-
-export interface InitCharge {
-  method: JsonRpcMethods.Init;
-  params: {
-    referenceId: string;
-    sender: {
-      accountAddress: string;
-      payerData: KYC;
-    };
-    transactionHash: string;
-  };
-}
